+++ lib/kafka_ex/new/client.ex
@@ -235,6 +235,17 @@ defmodule KafkaEx.New.Client do
     end
   end
 
+  def handle_call({:sync_group, consumer_group, generation_id, member_id, opts}, _from, state) do
+    if KafkaEx.valid_consumer_group?(consumer_group) and is_binary(consumer_group) do
+      case sync_group_request(consumer_group, generation_id, member_id, opts, state) do
+        {:error, error} -> {:reply, {:error, error}, state}
+        {result, updated_state} -> {:reply, result, updated_state}
+      end
+    else
+      {:reply, {:error, :invalid_consumer_group}, state}
+    end
+  end
+
   def handle_call({:kayrock_request, request, node_selector}, _from, state) do
     {response, updated_state} = kayrock_network_request(request, node_selector, state)
 
@@ -371,6 +382,22 @@ defmodule KafkaEx.New.Client do
     end
   end
 
+  defp sync_group_request(consumer_group, generation_id, member_id, opts, state) do
+    node_selector = NodeSelector.consumer_group(consumer_group)
+
+    req_data = [
+      {:group_id, consumer_group},
+      {:generation_id, generation_id},
+      {:member_id, member_id}
+      | opts
+    ]
+
+    case RequestBuilder.sync_group_request(req_data, state) do
+      {:ok, request} -> handle_sync_group_request(request, node_selector, state)
+      {:error, error} -> {:error, error}
+    end
+  end
+
   # ----------------------------------------------------------------------------------------------------
   defp handle_describe_group_request(request, node_selector, state) do
     handle_request_with_retry(request, &ResponseParser.describe_groups_response/1, node_selector, state)
@@ -394,6 +421,10 @@ defmodule KafkaEx.New.Client do
 
   defp handle_leave_group_request(request, node_selector, state) do
     handle_request_with_retry(request, &ResponseParser.leave_group_response/1, node_selector, state)
+  end
+
+  defp handle_sync_group_request(request, node_selector, state) do
+    handle_request_with_retry(request, &ResponseParser.sync_group_response/1, node_selector, state)
   end
 
   # ----------------------------------------------------------------------------------------------------