+++ lib/kafka_ex/new/protocols/kayrock/sync_group/v0_request_impl.ex
@@ -1,0 +1,11 @@
+defimpl KafkaEx.New.Protocols.Kayrock.SyncGroup.Request, for: Kayrock.SyncGroup.V0.Request do
+  @moduledoc """
+  Implementation for SyncGroup v0 Request.
+  """
+
+  alias KafkaEx.New.Protocols.Kayrock.SyncGroup.RequestHelpers
+
+  def build_request(request_template, opts) do
+    RequestHelpers.build_request_from_template(request_template, opts)
+  end
+end